Yank shared context to spec support file
