+++ spec/lib/everything/blog/output_spec.rb
@@ -1,26 +1,10 @@
 require 'pp' # Helps prevent an error like: 'superclass mismatch for class File'
 require 'bundler/setup'
 Bundler.require(:default)
+require './spec/support/shared'
 require './lib/everything/blog/output'
 
 describe Everything::Blog::Output do
-  shared_context 'stub out blog output path' do
-    let(:fake_blog_output_path) do
-      '/fake/blog/output'
-    end
-    let(:expected_absolute_path) do
-      fake_blog_output_path
-    end
-
-    before do
-      without_partial_double_verification do
-        allow(Fastenv)
-          .to receive(:blog_output_path)
-          .and_return(fake_blog_output_path)
-      end
-    end
-  end
-
   describe '.absolute_path' do
     include_context 'stub out blog output path'
