+++ spec/support/shared.rb
@@ -82,3 +82,20 @@ shared_context 'with deleted metadata file' do
   end
 end
 
+shared_context 'stub out blog output path' do
+  let(:fake_blog_output_path) do
+    '/fake/blog/output'
+  end
+  let(:expected_absolute_path) do
+    fake_blog_output_path
+  end
+
+  before do
+    without_partial_double_verification do
+      allow(Fastenv)
+        .to receive(:blog_output_path)
+        .and_return(fake_blog_output_path)
+    end
+  end
+end
+
